[BUGFIX] Fix frameset module property for FileList
BackendController::generateJavascript() creates the
top.fsMod object to provide frameset modules an easy
way of storing their state. The FileList module e.g.
stores the current folder identifier. Therefore, after
switching back to the FileList module, the previously
selected folder is used.

This behaviour is now restored by replacing the
"instanceof" check, which always evaluated to FALSE,
with a typeof check. It's furthermore now ensured
that the id will always be set, even if recentIds
is not yet an object.

Resolves: #92811
Releases: master, 10.4
Change-Id: I36227775bfadfca4d43a719b1508f78976b6e342
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67659
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>