[TASK] Drop unneeded renderElement from lit-helper
renderElement() was added in patchset 2 of #93446.
The code was then refactored to use renderHTML, but the method was
forgotten to be removed in the final patchset.

Now, the intention of the mentioned patchset 2 was to render nodes
directly to the DOM in PageTreeDragHandler, instead of rendering to
a temporary element, just to export and re-import as HTML string
with jQuery/innerHTML.

Therefore PageTreeDragHandler is now refactored to a new,
(internal) method renderNodes() which allows the returned
nodes to be directly attached to the document.

Also, while at it, all other lit-helper methods are marked as
internal as well, as they are still subject to change:
 * `renderHTML()` is only needed when mixing traditional jQuery
   rendering and lit-html (same as for the new `renderNodes()`.)
 * The `icon()` method could likely be replaced by a custom
   <typo3-backend-icon> component
 * `lll()` is only of use for globally available language identifiers
   (which are added via inline JS)  So this is also likely to change
   when more backend modules start to use lit-html/LitElement.

Resolves: #93464
Related: #93446
Releases: master
Change-Id: I61ab778e28e32b75b6e7965b75259beedbb10ae3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67663
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>