[FEATURE] Enable multi-level language fallback for rendered content
This change enables functionality of every content
to use the fallback chain to loop over each fallback language
to check for a fallback content in a different language ID.

This is now in line with the page resolving fallback handling
and finally completes the options within site configuration
"fallbackChain".

This works in Extbase and TypoScript environments.

Resolves: #88137
Releases: main
Change-Id: I987a296c6e483967d497bf59e71b8c5cbcd54938
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67893
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>