[BUGFIX] Avoid SQL error in indexed_search statistics
When an entered search word longer than 50 characters
was used in indexed_search previously, a TYPO3 Exception
was thrown due to a SQL error when having search statistics
enabled.

This change modifies the added statistic and cuts down
the search stats to a maximum length of 50 characters
as the DB field has only 50 characters.

Resolves: #93385
Related: #92993
Releases: master, 10.4
Change-Id: Iea691446571c31378c79c7ac27a7fe98b0fa458a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67926
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>