[BUGFIX] Fix margins of form labels
Labels within FormEngine are bold, but due to Boostrap 5
all labels in all places were bold and created margin, making
the rendering of the rest of the TYPO3 Backend hard-to-grasp,
as well as FormEngine not allowing to have a different view
for labels of fields, and labels of actual form elements.

This change adapts the margin and font-weight accordingly.

Resolves: #93519
Releases: master
Change-Id: I3bf0fda432cf43309fe3909a22784609ddeaa904
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67952
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>