[TASK] Do not require sudo mode in development context
The sudo mode, introduced in #92836, is required whenever
the install tool is accessed in the backend.

Especially in development context accessing the install
tool is a frequent task, e.g. for clearing all caches,
to test global configuration or to run the database
analyzer.

Therefore, the sudo mode is now not longer required
while the installations application context is set
to "Development".

Resolves: #93160
Releases: master, 10.4, 9.5
Change-Id: If61fa08847181491c01417d301a6bc1f480bae1b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68094
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>