[BUGFIX] Set request object in EXT:form EmailFinisher
Because the request object was not set for the FluidEmail
variant of the EXT:form EmailFinisher, the TYPO3 logo as
well as the website URL in the footer were not displayed
when using the SystemEmail layout from Core.

This is now fixed by setting the request, if available.

Resolves: #93588
Releases: master, 10.4
Change-Id: Ie6450b1a7a5728f1384e8010f02868bc26479c32
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68115
Tested-by: core-ci <typo3@b13.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>