[TASK] Deal with xdebug 3
Adapt testing related docker-compose.yml to properly
work with xdebug 3.0. This is packaged in the testing
container images already and is the only version
compatible with PHP 8.

To spice up developers life, xdebug 3.0 drops important
config options in a non b/w compatible way, introduces
new ones, where some of them overwrite others, and does
further stuff like changing the default step debugger
port for no apparent reason.

Since xdebug can be such a pain, runTests.sh comes with
-x option, so it (hopefully) 'just works' for devs using
the script in testing context. The setup of course breaks
with xdebug 3.0 and is fixed with the patch.

As single advantage, it is now easier to disable xdebug,
so the 'prepared php.ini file without xdebug' hack in
the test container images can be dropped.

Change-Id: Id5810a5a6d988bfe5e814a6781d12d93baa97972
Resolves: #93599
Releases: master, 10.4, 9.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68137
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>