[BUGFIX] Skip MFA in switch-user mode
In case of switching to a user, having MFA enabled,
with a user, having MFA disabled, the switching user
is required to pass the target users' MFA, because the
transformed session does not contain the `mfa` key.

Since it's obviously not possible to pass another users'
MFA, we need some exception for such scenario.

Therefore, the evaluateMfaRequirements() method
from AbstractUserAuthentication is overwritten by
BackendUserAuthentication to check if the current
session is a switch-user session. In this case,
MFA is skipped. Otherwise the parent method is
executed. Skipping MFA will be logged.

Resolves: #93624
Releases: master
Change-Id: I0625b23bee5aa202a20e2bb7ad46b6e606769134
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68164
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>