[BUGFIX] Inject Logger into LanguagePackService
The EXT:install ServiceProvider introduced in #89892 didn't respect
that LanguagePackService implements LoggerAwareInterface. That means
the logger variable was null and usages of the logger resulted in
exceptions.

Refactor the class to require the logger instance to always be
injected via constructor injection, to prevent this from happening
again.

Resolves: #93630
Related: #89892
Releases: master, 10.4
Change-Id: I71309863e096e5289336831ec4da294a5bcb65cf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68172
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>