[TASK] Show otpauth URL next to shared secret
The MFA TOTP provider is based on a shared secret. Therefore,
to set this provider up, the secret has to be exchanged between
TYPO3 and the corresponding OTP application or device.

This can either be done by:

* Scanning the QR code
* Entering the shared secret directly

While the shared secret does not contain any additional data,
the QR code contains the so-called otpauth URL, a common URI
scheme, accepted by most OTP applications. This URL features
additional information like the issuer and the related account.

To improve the usability, the user is now able to also access
this URL directly through a new modal with further explanation.

Resolves: #93666
Releases: master
Change-Id: I27dc805bcca7aeb29d15a5c20223426999277625
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68293
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>