[BUGFIX] Adjust tab menu styling
Because bootstrap 5 requires the `.nav-link` class for styling
of nav tab links, our custom SCSS now uses the class selector
together with the type selector. Otherwise it will be overruled
by the default bootstrap styles because of their specificity.

See: https://getbootstrap.com/docs/5.0/components/navs-tabs/#fill-and-justify

Furthermore is the `.active` class, indicating the active tab,
not longer set on the `.nav-item` element, but on the `.nav-link`
element. The custom SCSS is therefore adjusted to respect this
changed behaviour.

Resolves: #93691
Releases: master
Change-Id: I57260c370a12dd997652065fa6eebb82705eec38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68302
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>