[TASK] Separate concerns in SVGTree implementations
This patch provides the following changes in order to
have better abstractions and encapsulate the dependencies
of each implementation.

1. SelectTree (Category Tree)

The TreeToolbar is now clearly coupled into the SelectTree
and built as LitElement inside the SelectTreeElement.

SelectTree now contains all relevant functionality
regarding selecting multiple items, and shows settings
(SelectTreeSettings) which were only used in SelectTree.

This is especially relevant for the multiple item selection,
and the exclusiveKey handling.

2. PageTree + FileStorageTree
Obsolete public methods are removed, and visibility is changed.
Filtering is now separated in favor of a non-public property.
By using Types (in e.g. PageTreeDragDrop) the visibility is also adapted
Obsolete wrapper methods are removed.

Resolves: #93692
Releases: master
Change-Id: Ifead098bb80e37d5c7ad5e1d0eabda14aa7087ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68303
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>