[TASK] Clean up ignored phpstan errors
Remove entries from phpstan.neon for ignored errors
that have been fixed meanwhile. Parameter
reportUnmatchedIgnoredErrors finds those, but it is
false by default. We may drop that parameter with a
dedicated patch, but are currently not sure why it
is commented as 'needed for bamboo' at the moment.

Resolves: #93711
Releases: master
Change-Id: I9439b2144f60f0a4021b307ef5bacaa23801880e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68358
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>