[TASK] Stabilize acceptance tests
The install tool login is directly checking if the text
"The Install Tool is locked" is shown. Since this text
comes from an async request, this might take some time.

Also wait for the page to reload after creating the
ENABLE_INSTALL_TOOL file before trying to fill the
password field.

Resolves: #93715
Releases: master
Change-Id: I03230ee8107afe24c5c61d762f59ba0658e533d5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68361
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>