[TASK] Use Lit-based template for SVG tree wrapper
The SVG tree is now building the SVG elements
and needed containers ("<g>")  via lit-helper.

In addition, the update() method is renamed to
this.updateVisibleNodes() to easier integrate
this as Lit element, as update() is a reserved protected
method in lit.

The initialize() method is much cleaner, as the event listener
registration is now separated in a custom method.

Resolves: #93724
Releases: master
Change-Id: I945cd620f9900c6ea535dfe5f0d44ee5ced46f89
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68367
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>