[TASK] Drop TsTemplate 'breakpoint' handling
The backend template analyzer module has a 'breakpoint'
handling that works as follows: When a template is shown and
line numbers are enabled and one line number is clicked, a
link parameter jumps to the template object browser and
instructs it to parse TypoScript only up to this line to show
the state of constants or setup array up until this parsed point.

This hidden and not documented feature has been around since
'ever', with only a few persons knowing, understanding or even
actively using it. Last patch in this area was for TYPO3 4.4.

The implementation is quite messy and goes through various
classes, a bunch of properties and some inline javascript.

The patch drops the feature: It is not well known and of
limited use - the frontend does not reflect it and TypoScript
in general tends to become less complex nowadays.

Resolves: #93726
Related: #23265
Related: #23246
Related: #55761
Releases: master
Change-Id: I4f0c7f507a5a6e212f166206b5d3e606071b46c4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68371
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>