[BUGFIX] Remove DeferredBackendImageProcessor from configuration
Remove DeferredBackendImageProcessor from default configuraion,
because it requires usage of a newly introduced API for
third party extensions to properly function in some cases,
which factually is a breaking change.

The code still remains there, so that this feature can be activated
again for installations, where these cases do not matter or can be handled
correctly.

For TYPO3 11.x the configuration will change, but behaviour will be
changed to have less impact and cover 99% of the use cases.
This will be done in a separate change.

Releases: 10.4
Resolves: #93766
Resolves: #93133
Resolves: #93245
Change-Id: I9b2c3e4a8e7c8a3da567c2eefa25687194a59fe7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68518
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>