[BUGFIX] Fix workspace pagination
Since the workspace pagination took also child records
into account, which are however never being displayed
without their parent, the view broke as soon as such child
record would have been the first item on the next or the last
item on the previous page.

Therefore, the "total" count, used by the JavaScript pagination
to calculate the amount of pages, does not longer take any
child records into account.

The data array, containing the records for the requested page,
is now filled until the limit of records is reached with only parent
records. Therefore, the actual record items count can differ from
the proposed limit, since the child records are added without
increasing the corresponding counter.

The same functionality does also take place on calculation of the
offset value. This means, the offset provided by the pagination is
e.g. 30, but the backend side offset must be 30 + the amount of
child records in the previous 30 parent records.

Another special case is now also handled properly: When the last
record to be added is a parent record, having child records, they
need to be fetched recursively as well. Otherwise this would lead
to the same bug as before, since such record would then be the
first record on the next page, which does not work.

Resolves: #93645
Releases: master, 10.4
Change-Id: I742486fc2bf2a47d6a4f19bd4fae51b21d498074
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68519
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>