[BUGFIX] Fix TypeError in PreviewUriBuilder
With #91123 a new PreviewUriBuilder class was introduced.
This class features a factory method `create`, with a type
hinted parameter signature. This now leads to a TypeError
in EXT:workspaces, e.g. when importing datasets in tests.

This is fixed by typecasting the variable to be passed
to the factory method.

Resolves: #93800
Releases: master
Change-Id: I7b1f567514da53bad0d16ac49fdfc9a33c2b7547
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68558
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>