[BUGFIX] Ensure compared values are of same type
When strict type comparison was introduced to
the OptionViewHelper in #90540, it was not properly
ensured that the values to compare are of same type.

The issue is fixed by always typecasting the selected
value to string, before comparing it to the value in
question, which will always be a string, according to
its type hint.

Resolves: #93801
Releases: master, 10.4
Change-Id: I3ea1d885a35e3223919facfe6d1825332914a7ed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68592
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>