[BUGFIX] Fix descriptions of selectCheckBox items
With #91008 the grouping and sorting feature was
introduced to the TCA select type. As a consequence,
the select items array key for the help text changed
from `3` to `4`.

The SelectCheckBox now correctly checks the `4`
key for the presence of a description.

In v11 the absence of a title value, which is common
in this scenario, did furthermore lead to TypeError
on rendering the popover. This is fixed by assigning
an empty string as default value on initialization of
the popover options. Otherwise this option would be
initialized as 'undefined' and then trigger the
TypeError on sanitizing.

Additionally the TCA properties special=table,
special=exclude and special=custom are adjusted to
add the correct value as help text to the items array.
The value is now added directly as string, as there
is no need to define the array while only providing
the description.

As a drive-by-change, the itemsProcFuncs of the
dashboard and the MFA component do now also add
the description for their access list field. This
is done because the dashboard initially defined the
description already but had to remove it again due
to this bug, see: #91152. The MFA component also
just hadn't added it because the bug hadn't been
resolved at that time.

Resolves: #93331
Resolves: #92383
Releases: master, 10.4
Change-Id: Ie96bcabea88b377490c24b0e60cfca8337e9ec52
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68619
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>