[TASK] Deprecate special property of TCA type select
In #57082 the special=languages option was migrated to the
new TCA type "language". The remaining options for "special"
are mainly used for be_users and be_groups only and can
also be implemented as `itemsProcFunc` functions. This reduces
unnecessary complexity for the TCA "select" type.

Therefore, this patch migrates all remaining options of the
"special" property to dedicated itemsProcFunc's and deprecates
the use of this property along with the corresponding
`addItemsFromSpecial` method in AbstractItemProvider.

Since the remaining options are mainly used internally, the new
class, containing the itemsProcFunc's, is marked as internal.

Resolves: #93837
Releases: master
Change-Id: I32be195c70babf387e1f1b42cbea03f6093dcefb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68622
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>