Update comment to remove never used typo3_conf_var "transport_smtp_port"
remove the comment with transport_smtp_port as it was never used within typo3 core. Introduced way back in https://forge.typo3.org/attachments/15596, but already documented that the port has to be written within transport_smtp_server with a colon. Update the line of transport_smtp_server to a proper example by adding colon an port.