[TASK] Optimize search box for record lists
The search box is currently available in
* Page module (docheader)
* List module (docheader)
* Record Selector / Element Browser

All implementations are different but use the same code base:
* The outer wrapper functionality is only needed for the doc header
* The page module is not using the Fluid-based snippet

Additionally, styling is broken due to Bootstrap 5 upgrade.

The change reduces CSS and JS initialization by consolidating
the rendering of the search box and removing the buggy showLimit
functionality since there is no pagination the placeholder is not used.

Ideally, this code would be separated into a separate class
as it is only related to a fragment of DatabaseRecordList.
This is to be handled with a separate patch.

Resolves: #93849
Releases: master
Change-Id: I20e406f0ed5bf74503fb45c9ad93cf0e6a91f153
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68667
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>