[TASK] Clean up TypeScript code for Element Browser / Link Picker
The Link Picker (RecordLinkHandler) and Element Browser logic
has been simplified to avoid jQuery where possible and to reduce
HTML rendering.

The data-icon information is never evaluated in TypeScript/JavaScript
and can be removed from any kind of HTML rendering, making the
transferred amount of HTML much smaller.

In addition, the insertElement method has been cleaned up to only
use the properties that are evaluated.

As a small side-effect, the InlineControlContainer does not
throw a console.warn if a post message cannot be evaluated
(which is typically the case for non-inline references).

Resolves: #93850
Releases: master
Change-Id: Ie850a3520ba5b448f5a1d0a42aa656b972b8d3a6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68668
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>