[BUGFIX] Prevent saveAndClose in MultiStep wizard
With #90461 a new option "saveAndClose" was added
to the newContentElement wizard. When opening the
wizard and selecting a content type with such option set,
the new element will directly be created - no FormEngine
interaction - and placed at the position (colPos, sorting),
the user opened the wizard.

This functionality however does not work when the wizard
was opened without such page context, for example via the
"new record view" in the recordlist. The position must then
be defined by the user explicitly in the MultiStep wizard, so
no "saveAndClose" functionality is available in such cases.

Resolves: #93876
Releases: master, 10.4
Change-Id: I8c7c7b632533a6f2d2bbe01a103a25ce441acf2c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68812
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>