[TASK] Unify authentication form styles
Using backend extension configuration, it's possible
to customize the authentication forms, e.g. login and
password recovery.

Since MFA was added as another part of this process,
the corresponding authentication forms must also
respect the customized styling.

This is achieved by extracting the logic form
LoginController and adding it to a dedicated
class `AuthenticationStyleInformation`.

Resolves: #93898
Releases: master
Change-Id: Ia0209017592ac9b168934df3ed3ab8bc08a1913d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68765
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>