[BUGFIX] Prematurely end data array processing on invalid item
In case the amount of items, shown in the workspace module,
does not exceed the pagination limit - default 30 - the items
array was previously filled with invalid items till the limit
was reached.

This is now fixed by a guard clause which prematurely ends the
processing (the corresponding loop) as soon as no more item
exists.

This guard clause is also implemented in the calculation of
the start value. Usually there shouldn't be any invalid items,
but as such behaviour can not always be ruled out, especially
on multi-user systems, it seems reasonable to add it here, too.

Resolves: #93915
Relates: #93645
Releases: master, 10.4
Change-Id: I619c9063f12c3b1d5c446f9ca1a0e35e646f90f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68793
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>