[BUGFIX] Display correct message if slug suggestion is modified
If a user enters a slug proposal, which is already taken by
another page in the same site, the automatically modified
slug is now shown along with the corresponding message.

The actual functionality already existed, but was never used
due to a Bug in the AJAX response.

In addition, the label for the warning message does not have
the base / slug at the very end, so sprintf() must be used
AFTER htmlspecialchars().

Resolves: #93953
Releases: master, 10.4
Change-Id: I395c39788ff25fc1b8cefe78e07b609a4be36e2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68852
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>