[TASK] Clean up legacy tree implementations
Various code occurrences within the legacy tree implementations,
which were marked as internal are removed
(e.g. the method to evaluate a "PM" query parameter)
as the functionality is not needed anymore, along
with some unused properties as well, which are never evaluated
anymore due to the removal of the methods.

Resolves: #93990
Releases: master
Change-Id: I3b8841959156f87fafed7eb58bea01770ea9dce6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68877
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>