[BUGFIX] Use correct route path for submodule request
The path for the Belog module changed in v11 and does
now contain the `/module` prefix. Since it's obviously
never a good idea to use such hard-coded values, as can
be seen in this example, the route path is now generated
by the UriBuilder. This will prevent further bugfixes
in case the path structure will change again.

Additionally, the $GLOBALS['TYPO3_REQUEST'] declaration
is removed, since #93853 was already merged.

Resolves: #94018
Releases: master
Change-Id: I260357370da0b7a6bd7b9e8ea6f5ea0c8503f9ea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68911
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>