[BUGFIX] Apply TCAdefaults to DataHandler if FormEngine was not used before
If TCAdefaults.pages.categories = 4 is used in userTSconfig, and a new
page is created via Drag&Drop or via "Create Multiple New Pages" via
ClickMenu, TCAdefaults are now added to the incomingFieldArray and
not just the fieldArray.

This is also great for translations (using "Just Localized"(tm) technology),
where the defaults are respected as well now.

Resolves: #92146
Resolves: #77847
Related: #89211
Releases: master, 10.4
Change-Id: I87a2f2d486831f9bda9a979c4ff99d4e8b2ada2f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68938
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>