[BUGFIX] Add active class to correct element in EXT:workspaces
The tablist inside the element information modal
in the workspaces module did not longer set the
active state on the correct tab item.

This is due to a change in bootstrap 5, since an
active tab is highlighted by setting the `active`
class on the tab item link (`nav-link`) instead of
the tab item (`nav-item`) itself.

This is fixed by adding the `active` class to the
correct element again.

Further reference:
https://getbootstrap.com/docs/5.0/components/navs-tabs

Resolves: #94028
Related: #93119
Releases: master
Change-Id: I21ee8ede69876feef047c6303c4fbe890f898a65
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68928
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>