[BUGFIX] Keep unmapped ?type parameter when using PageTypeDecorator
This change keeps the type parameter (e.g. ?type=13) even
if PageTypeDecorator is used when building URLs but "13" is not
part of the map in the site configuration.

A test is added in order to make sure this functionality
will not break (again), see
https://review.typo3.org/c/Packages/TYPO3.CMS/+/62383
for the original fix without tests.

Resolves: #87104
Related: #87817
Related: #88836
Releases: master, 10.4, 9.5
Change-Id: Ic7f82bfa9f28f971162e1af1b557188f61446462
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69115
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>