[FEATURE] Add Shona (sn) as supported language
Shona is a Bantu language of the Shona people of Zimbabwe.
It is one of the most widely spoken Bantu languages.

Also see https://en.wikipedia.org/wiki/Shona_language

This change adds support for Shona to be selected as TYPO3 Language
and to connect to TYPO3 Localizations (XLF files, Site Configuration)
by default.

Resolves: #93114
Releases: master
Change-Id: Id25823fc5021d9cc3480fb57627fe3cfcfbcb614
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68987
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>