[TASK] Streamline EXT:belog filter
To be more consistent with the backend user and redirects
module, the filter section of the backend log module is
streamlined.
The clumsy time range drop down is replaced, the 'user defined'
time selection via date pickers is always shown instead.
The hard to understand 'group by page' when called from
within info module having a depth > 0 is dropped, since a
similar information can be achieved more naturally using
the page tree selection.
The filter drop-downs don't reload automatically anymore
to be in-line with other modules, 'set' is renamed to
'Filter' and a 'Reset' button is added.

Resolves: #94072
Releases: master
Change-Id: Ia0a2229bc8f37a717a0e10a39e7337d67a32456c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69036
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>