[BUGFIX] Fix InputLinkElement to be visible again
Since the bootstrap v5 update, the inputLinkElement is styled with flex
form instead of display table. This became a problem, because this broke
the appearance of the explanation toggling.

The initial idea to hide the complete clearable wrap didn't work
properly as it is loaded dynamically and it's not ensured it exists
when trying to hide it.

To circumvent this problem, the styling is now done with some extra css.
Additionally the wrong form-control class is removed from the clearable
wrap and the styling is now also moved to own css.

Resolves: #93941
Releases: master
Change-Id: I0bb0d1cf26ed40dd3c87a58e08cc39e449b96c0d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69064
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>